refactor(brett): extract phases.ts (transitionPhase, buildStateFromMutations)
